requirements.txt
# Core Libraries
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
matplotlib>=3.4.0
tqdm>=4.60.0
argparse>=1.1
scikit-learn>=1.0.0
pickle5>=0.0.12
jsonschema>=3.2.0
csvkit>=1.0.6

# Lightning (Fabric/Trainer)
lightning>=2.0.0
torchmetrics>=0.11.0

# Custom Dependencies
clip @ git+https://github.com/openai/CLIP.git
sklearn>=0.24.2

# Optional dependencies for additional data loaders and utilities
pillow>=8.0.0
